+++ tests/ember/ops/test_sub.cpp
@@ -8,48 +8,48 @@
 using namespace ember;
 
 TEST(TensorSubtraction, ScalarTensorsCanBeSubtracted) {
-  Tensor a = {1.0f};
-  Tensor b = {5.0f};
+  Tensor a = {1.0};
+  Tensor b = {5.0};
   Tensor c = a - b;
 
   auto difference = c.data;
-  auto expected_difference = xt::xarray<float>({-4.0f});
+  auto expected_difference = xt::xarray<double>({-4.0});
   EXPECT_TRUE(xt::allclose(difference, expected_difference));
 
   c.backward();
 
-  EXPECT_TRUE(xt::allclose(a.gradient->data, xt::xarray<float>{1.0f}));
-  EXPECT_TRUE(xt::allclose(b.gradient->data, xt::xarray<float>{-1.0f}));
+  EXPECT_TRUE(xt::allclose(a.gradient->data, xt::xarray<double>{1.0}));
+  EXPECT_TRUE(xt::allclose(b.gradient->data, xt::xarray<double>{-1.0}));
 }
 
 TEST(TensorSubtraction, MultidimensionalTensorsCanBeSubtracted) {
-  Tensor a = {{1.0f, 9.0f}, {3.0f, 2.2f}};
-  Tensor b = {{5.0f, 3.0f}, {2.0f, 1.3f}};
+  Tensor a = {{1.0, 9.0}, {3.0, 2.2}};
+  Tensor b = {{5.0, 3.0}, {2.0, 1.3}};
   Tensor c = a - b;
 
-  xt::xarray<float> difference = c.data;
-  xt::xarray<float> expected_difference = {{-4.0f, 6.0f}, {1.0f, 0.9f}};
+  xt::xarray<double> difference = c.data;
+  xt::xarray<double> expected_difference = {{-4.0, 6.0}, {1.0, 0.9}};
 
   EXPECT_TRUE(xt::allclose(difference, expected_difference));
 
   c.backward();
 
-  EXPECT_TRUE(xt::allclose(a.gradient->data, xt::xarray<float>{{1.0f, 1.0f}, {1.0f, 1.0f}}));
-  EXPECT_TRUE(xt::allclose(b.gradient->data, xt::xarray<float>{{-1.0f, -1.0f}, {-1.0f, -1.0f}}));
+  EXPECT_TRUE(xt::allclose(a.gradient->data, xt::xarray<double>{{1.0, 1.0}, {1.0, 1.0}}));
+  EXPECT_TRUE(xt::allclose(b.gradient->data, xt::xarray<double>{{-1.0, -1.0}, {-1.0, -1.0}}));
 }
 
 TEST(TensorSubtraction, AnonymousIntermediateTensorsCanBeSubtracted) {
-  Tensor a = {{7.0f, 3.0f}, {4.0f, 1.0f}};
-  Tensor b = {{8.0f, 2.0f}, {5.0f, 0.0f}};
+  Tensor a = {{7.0, 3.0}, {4.0, 1.0}};
+  Tensor b = {{8.0, 2.0}, {5.0, 0.0}};
   Tensor c = a - b;
   
-  Tensor d = (c - Tensor({{3.0f, 3.0f}, {3.0f, 3.0f}})) - 
-             (c - Tensor({{5.0f, 5.0f}, {5.0f, 5.0f}}));
-  EXPECT_TRUE(xt::allclose(d.data, xt::xarray<float>({{2.0f, 2.0f}, {2.0f, 2.0f}})));
+  Tensor d = (c - Tensor({{3.0, 3.0}, {3.0, 3.0}})) - 
+             (c - Tensor({{5.0, 5.0}, {5.0, 5.0}}));
+  EXPECT_TRUE(xt::allclose(d.data, xt::xarray<double>({{2.0, 2.0}, {2.0, 2.0}})));
 
   d.backward();
-  EXPECT_TRUE(xt::allclose(a.gradient->data, xt::xarray<float>({{0.0f, 0.0f}, {0.0f, 0.0f}})));
-  EXPECT_TRUE(xt::allclose(b.gradient->data, xt::xarray<float>({{0.0f, 0.0f}, {0.0f, 0.0f}})));
+  EXPECT_TRUE(xt::allclose(a.gradient->data, xt::xarray<double>({{0.0, 0.0}, {0.0, 0.0}})));
+  EXPECT_TRUE(xt::allclose(b.gradient->data, xt::xarray<double>({{0.0, 0.0}, {0.0, 0.0}})));
 }
 
 TEST(TensorSubtraction, BroadcastingWorks) {
@@ -65,26 +65,26 @@ TEST(TensorSubtraction, BroadcastingWorks) {
 }
 
 TEST(TensorSubtraction, ZeroSubtractionPreservesGradients) {
-    Tensor a = {1.0f, 2.0f, 3.0f};
-    Tensor b = {1.0f, 2.0f, 3.0f};
+    Tensor a = {1.0, 2.0, 3.0};
+    Tensor b = {1.0, 2.0, 3.0};
     Tensor c = a - b;
 
-    EXPECT_TRUE(xt::allclose(c.data, xt::xarray<float>{0.0f, 0.0f, 0.0f}));
+    EXPECT_TRUE(xt::allclose(c.data, xt::xarray<double>{0.0, 0.0, 0.0}));
 
     c.backward();
-    EXPECT_TRUE(xt::allclose(a.gradient->data, xt::xarray<float>{1.0f, 1.0f, 1.0f}));
-    EXPECT_TRUE(xt::allclose(b.gradient->data, xt::xarray<float>{-1.0f, -1.0f, -1.0f}));
+    EXPECT_TRUE(xt::allclose(a.gradient->data, xt::xarray<double>{1.0, 1.0, 1.0}));
+    EXPECT_TRUE(xt::allclose(b.gradient->data, xt::xarray<double>{-1.0, -1.0, -1.0}));
 }
 
 TEST(TensorSubtraction, GradientWithBroadcastAndScalar) {
-    Tensor matrix = {{1.0f, 2.0f}, {3.0f, 4.0f}};
-    Tensor scalar = {2.0f};
+    Tensor matrix = {{1.0, 2.0}, {3.0, 4.0}};
+    Tensor scalar = {2.0};
     
     auto c = matrix - scalar;
     c.backward();
 
     // Expected: ∂(m-s)/∂m = 1, ∂(m-s)/∂s = -sum(1)
-    EXPECT_TRUE(xt::allclose(matrix.gradient->data, xt::ones_like(matrix.data), 0.001f));
-    EXPECT_TRUE(xt::allclose(scalar.gradient->data, xt::xarray<float>({-4.0f}), 0.001f));
+    EXPECT_TRUE(xt::allclose(matrix.gradient->data, xt::ones_like(matrix.data), 0.001));
+    EXPECT_TRUE(xt::allclose(scalar.gradient->data, xt::xarray<double>({-4.0}), 0.001));
 }
